fix(preprocessing): prevent int32 overflow in Otsu histogram sums and between-class variance
CodeQL cs/loss-of-precision (alerts 13390-13392), Binarization.cs:105-125: `wB * wF` overflows
at about 92.7k pixels (a ~305x305 image), so Otsu never found a threshold and returned 0
(every pixel foreground). The histogram weighted sums overflow for large pages as well.

+++ src/Preprocessing/Document/Binarization.cs
@@ -102,7 +102,7 @@ public class Binarization<T> : IDisposable
         int total = gray.Length;
         double sum = 0;
         for (int i = 0; i < 256; i++)
-            sum += i * histogram[i];
+            sum += (double)i * histogram[i];
 
         double sumB = 0;
         int wB = 0;
@@ -117,12 +117,12 @@ public class Binarization<T> : IDisposable
             int wF = total - wB;
             if (wF == 0) break;
 
-            sumB += t * histogram[t];
+            sumB += (double)t * histogram[t];
 
             double mB = sumB / wB;
             double mF = (sum - sumB) / wF;
 
-            double variance = wB * wF * (mB - mF) * (mB - mF);
+            double variance = (double)wB * wF * (mB - mF) * (mB - mF);
 
             if (variance > maxVariance)
             {
